File and code refactor

- API usage over 90%  notice
- Missing affiliate IDs notice
- No networks selected notice
- No merchants selected notice
- User account data helper functions
- Removed all functionality from Env file
- Moved functions and actions into appropriate files
- Fixed undefined index related to "adservice_mid".
- Removed old, unused action.
- Prevented the selection of Belboon, Effiliation and Partnerize networks if required keys were missing.
- Moved the searchform require to all pages, not just admin.
- Added helper functions for group IDs
- Added new function that gets affiliate ID by network ID.

+++ hooks/admin/ajax.php
@@ -1,25 +1,29 @@
 <?php
 
-// Exit if accessed directly
-if ( ! defined( 'ABSPATH' ) ) {
-	exit;
-}
-
-if ( is_admin() ) {
-	add_action( 'wp_ajax_dfrapi_delete_cached_api_data', 'dfrapi_delete_cached_api_data' );
-	add_action( 'wp_ajax_dfrapi_test_api_connection', 'dfrapi_test_api_connection' );
-}
+defined( 'ABSPATH' ) || exit;
 
 /**
  * Delete cached API data.
+ *
+ * We can only delete if user has API requests remaining.
+ * This is because we'll need to make 1 request to rebuild 'dfrapi_account'.
  */
 function dfrapi_delete_cached_api_data() {
+
 	check_ajax_referer( 'dfrapi_ajax_nonce', 'dfrapi_security' );
-	// Only delete if user has API requests remaining. This is because we'll need to make 1 request to rebuild 'dfrapi_account'.
+
+	if ( ! current_user_can( 'edit_posts' ) ) {
+		_e( 'You do not have permission to perform this action.', 'datafeedr-api' );
+		die;
+	}
+
 	$status = dfrapi_api_get_status();
+
 	if ( ! array_key_exists( 'dfrapi_api_error', $status ) ) {
+
 		delete_option( 'dfrapi_account' );
 		$transient_options = get_option( 'dfrapi_transient_whitelist' );
+
 		if ( ! empty( $transient_options ) ) {
 			foreach ( $transient_options as $name ) {
 				$use_cache = wp_using_ext_object_cache( false );
@@ -27,14 +31,18 @@ function dfrapi_delete_cached_api_data() {
 				wp_using_ext_object_cache( $use_cache );
 			}
 		}
+
 		// Update account status immediately in case there are not enough API
 		// requests remaining in order to do so later.
 		$status = dfrapi_api_get_status();
 		update_option( 'dfrapi_account', $status );
 	}
+
 	_e( 'Cached API data deleted successfully.', 'datafeedr-api' );
 	die;
 }
+
+add_action( 'wp_ajax_dfrapi_delete_cached_api_data', 'dfrapi_delete_cached_api_data' );
 
 /**
  * Test API Connection.
@@ -43,10 +51,15 @@ function dfrapi_test_api_connection() {
 
 	check_ajax_referer( 'dfrapi_ajax_nonce', 'dfrapi_security' );
 
+	if ( ! current_user_can( 'edit_posts' ) ) {
+		_e( 'You do not have permission to perform this action.', 'datafeedr-api' );
+		die;
+	}
+
 	$url      = "https://api.datafeedr.com";
 	$response = wp_remote_get( $url );
 	$code     = wp_remote_retrieve_response_code( $response );
-	$success  = $code == '200';
+	$success  = $code === 200;
 	$color    = $success ? 'green' : 'red';
 	$heading  = $success ? __( 'Success!', 'datafeedr-api' ) : __( 'Failed!', 'datafeedr-api' );
 	$message  = $success
@@ -57,3 +70,5 @@ function dfrapi_test_api_connection() {
 
 	die;
 }
+
+add_action( 'wp_ajax_dfrapi_test_api_connection', 'dfrapi_test_api_connection' );
